Bug: 261731544
Merged-In: Ie0613785ae34e9ef08c82b714e27ac16c4da363a

+++ core/ble_request_manager.cc
@@ -20,7 +20,6 @@
 #include "chre/platform/fatal_error.h"
 #include "chre/platform/log.h"
 #include "chre/util/nested_data_ptr.h"
-#include "chre/util/system/event_callbacks.h"
 
 namespace chre {
 
@@ -460,8 +459,8 @@ void BleRequestManager::logStateToBuffer(DebugDumpWrapper &debugDump) const {
     if (log.enable && log.compliesWithBleSetting) {
       debugDump.print(" mode=%" PRIu8 " reportDelayMs=%" PRIu32
                       " rssiThreshold=%" PRId8 " scanCount=%" PRIu8 "\n",
-                      static_cast<uint8_t>(log.mode), log.reportDelayMs,
-                      log.rssiThreshold, log.scanFilterCount);
+                      log.mode, log.reportDelayMs, log.rssiThreshold,
+                      log.scanFilterCount);
     } else if (log.enable) {
       debugDump.print(" request did not comply with BLE setting\n");
     }
